BaseCandidateViewを修正
